+++ app/src/main/java/me/maagk/johannes/virtualpeer/survey/question/SliderQuestion.kt
@@ -1,6 +1,6 @@
 package me.maagk.johannes.virtualpeer.survey.question
 
-class SliderQuestion(question: String, var min: Int, var max: Int) : Question(question) {
+class SliderQuestion(question: String, var min: Float, var max: Float) : Question(question) {
 
     override var answer: Any? = 0f
